review: extract card item value subcomponents
