+++ web-components/src/components/modal/TxModal.tsx
@@ -1,15 +1,15 @@
-import React, { useState } from 'react';
-import { Box, Collapse, Theme } from '@mui/material';
+import React from 'react';
+import { Box, Theme } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import { SxProps } from '@mui/system';
 
 import { truncate } from '../../utils/truncate';
 import OutlinedButton from '../buttons/OutlinedButton';
-import { TextButton } from '../buttons/TextButton';
 import Card from '../cards/Card';
 import { LinkItem } from '../footer/footer-new';
 import Modal, { RegenModalProps } from '../modal';
-import { Label, Subtitle, Title } from '../typography';
+import { Label, Title } from '../typography';
+import { CardItemValue, CardItemValueList } from './TxModal.CardItemValue';
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -42,7 +42,7 @@ interface LinkProps extends LinkItem {
   sx?: SxProps<Theme>;
 }
 
-type LinkComponentProp = React.FC<LinkProps>;
+export type LinkComponentProp = React.FC<LinkProps>;
 
 export interface TxModalProps extends RegenModalProps {
   onButtonClick: () => void;
@@ -55,88 +55,6 @@ export interface TxModalProps extends RegenModalProps {
   icon?: JSX.Element;
   title?: string;
 }
-
-interface CardItemValueProps {
-  value: ItemValue;
-  color?: string;
-  linkComponent: LinkComponentProp;
-}
-
-const CardItemValue = ({
-  value,
-  color,
-  linkComponent: LinkComponent,
-}: CardItemValueProps): JSX.Element => {
-  return (
-    <Subtitle
-      key={value.name}
-      size="lg"
-      mobileSize="sm"
-      color={color || 'info.dark'}
-    >
-      {value.icon && value.icon}
-      {value.url ? (
-        <LinkComponent
-          sx={{ color: 'secondary.main' }}
-          href={value.url}
-          target={value.url.startsWith('/') ? '_self' : '_blank'}
-        >
-          {value.name}
-        </LinkComponent>
-      ) : (
-        <>{value.name}</>
-      )}
-    </Subtitle>
-  );
-};
-
-interface CardItemValueListProps {
-  value: ItemValue[];
-  color?: string;
-  linkComponent: LinkComponentProp;
-}
-
-const CardItemValueList = (props: CardItemValueListProps): JSX.Element => {
-  const [expanded, setExpanded] = useState<boolean>(false);
-
-  if (props.value.length <= 2) {
-    return (
-      <>
-        {props.value.map(row => (
-          <CardItemValue {...props} value={row} />
-        ))}
-      </>
-    );
-  }
-
-  const alwaysVisible = props.value.slice(0, 2);
-  const underCollapse = props.value.slice(2);
-
-  return (
-    <>
-      {alwaysVisible.map(row => (
-        <CardItemValue {...props} value={row} />
-      ))}
-
-      <TextButton
-        textSize="xxs"
-        onClick={() => setExpanded(!expanded)}
-        sx={{
-          px: [0],
-          ':hover': { bgcolor: 'transparent !important' },
-        }}
-      >
-        {expanded ? '- see less' : '+ see more'}
-      </TextButton>
-
-      <Collapse in={expanded}>
-        {underCollapse.map(row => (
-          <CardItemValue {...props} value={row} />
-        ))}
-      </Collapse>
-    </>
-  );
-};
 
 interface CardItemProps extends Item {
   linkComponent: LinkComponentProp;
